Pull out interface for SubQueueSelectorCache persistence
